requirements.txt
opencv-python>=4.12,<5.0
Pillow>=11.0,<12.0
